prevented duplicate line items for one product in a cart

Former-commit-id: 01055ad362f0f45f027dae60420b27bd46c7590c

+++ models/Product.js
@@ -7,7 +7,7 @@ module.exports = function(sequelize, DataTypes) {
       subcategory: { type: DataTypes.STRING },
       name: { type: DataTypes.STRING },
       description: { type: DataTypes.STRING },
-      price: { type: DataTypes.DECIMAL },
+      price: { type: DataTypes.INTEGER },
       quantity: { type: DataTypes.INTEGER },
       url: { type: DataTypes.STRING }
     },
